feat(largestRetangle): largest retangle problem, by now only solves simple problems
